requirements.txt
flask-socketio
eventlet
websockets
websocket-client
tensorflow
opencv-python
numpy
serial
